refactoring

+++ application/src/main/java/com/kbe/application/models/calculatorAPI/VAT.java
@@ -4,9 +4,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class VAT {
-    private double vatPrice;
+
+    private BigDecimal vatPrice;
 }
